你说得对但是
<<<Tidal<<<
>>>Wave>>>

+++ src/main/java/com/science/gtnl/mixins/late/TwistSpaceTechnology/MixinMM_DimensionallyTranscendentMatterPlasmaForgePrototypeMK2.java
@@ -37,7 +37,7 @@ public abstract class MixinMM_DimensionallyTranscendentMatterPlasmaForgePrototyp
     MultiExecutionCoreMachineSupportAllModuleBase<MixinMM_DimensionallyTranscendentMatterPlasmaForgePrototypeMK2> {
 
     @Shadow
-    private double runningTime;
+    private long runningTime;
 
     @Shadow
     protected double fuelCostMultiplier;
